Editing Employee Entry

+++ IspDataDb/dbo/Tables/PhoneType.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[PhoneType] (
-    [IdNo]             TINYINT      IDENTITY (1, 1) NOT NULL,
+    [IdNo]             SMALLINT     IDENTITY (1, 1) NOT NULL,
     [PhoneTypeCode]    VARCHAR (5)  COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [PhoneTypeName]    VARCHAR (15) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [PhoneTypeNameAra] VARCHAR (15) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -7,4 +7,6 @@
     [DateTimeStamp]    ROWVERSION   NULL,
     CONSTRAINT [PK_PhoneType] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
